Move lockUser() and changePassword() to LoginPageBL. Remove all waiters. Move logout logic to AfterMethod. Remove duplicate code.
