+++ src/main/java/com/opencart/steps/LoginPageBL.java
@@ -44,7 +44,19 @@ public class LoginPageBL {
         Assert.assertEquals(successLoginPage.getMyAccountTitle().getText(), expectedMessage, "Incorrect page title");
     }
 
+    public void changePassword(String email) {
+        new MainPageBL().getHeaderPageBL()
+                .clickOnMyAccountButton()
+                .clickOnLoginButton()
+                .clickOnForgottenPassword()
+                .changePasswordWithURL(email)
+                .verifyChangePassword();
+    }
+
     public LoginPageBL lockUser(String email) {
+        new MainPageBL().getHeaderPageBL()
+                .clickOnMyAccountButton()
+                .clickOnLoginButton();
         do {
             loginUser(email, "wrong_password");
         } while (!lockMessage());
@@ -52,19 +64,16 @@ public class LoginPageBL {
     }
 
     public ForgottenPageBL clickOnForgottenPassword() {
-        loginPage.waitUntilFindElement(loginPage.getForgottenPasswordButton());
         loginPage.getForgottenPasswordButton().click();
         return new ForgottenPageBL();
     }
 
     private void inputEmail(String email) {
-        loginPage.waitUntilFindElement(loginPage.getEmailInput());
         loginPage.getEmailInput().clear();
         loginPage.getEmailInput().sendKeys(email);
     }
 
     private void inputPassword(String password) {
-        loginPage.waitUntilFindElement(loginPage.getPasswordInput());
         loginPage.getPasswordInput().clear();
         loginPage.getPasswordInput().sendKeys(password);
     }
@@ -74,13 +83,11 @@ public class LoginPageBL {
     }
 
     private boolean lockMessage() {
-        loginPage.waitUntilFindElement(loginPage.getAlertMessage());
         String lockMessage = "Warning: Your account has exceeded allowed number of login attempts. Please try again in 1 hour.";
         return loginPage.getAlertMessage().getText().equals(lockMessage);
     }
 
     public void verifyUserLogin() {
-        loginPage.waitUntilFindElement(successLoginPage.getMyAccountTitle());
         String expectedMessage = "My Account";
         Assert.assertEquals(successLoginPage.getMyAccountTitle().getText(), expectedMessage, "Incorrect page title");
     }
