+++ src/main/java/com/opencart/steps/RegisterPageBL.java
@@ -21,7 +21,7 @@ public class RegisterPageBL {
         RegisterModel registerModel = RegisterModelRepository.getRegisterModel();
         inputFirstName(registerModel.getFirstName());
         inputLastName(registerModel.getLastName());
-        inputEmail(RandomEmailUtil.getRandomEmail());
+        inputEmail(registerModel.getEmail());
         inputTelephone(registerModel.getTelephone());
         inputPassword(registerModel.getPassword());
         chooseSubscribe(1);
@@ -33,31 +33,26 @@ public class RegisterPageBL {
     }
 
     private void inputFirstName(String firstName) {
-        registerPage.waitUntilFindElement(registerPage.getFirstNameInput());
         registerPage.getFirstNameInput().clear();
         registerPage.getFirstNameInput().sendKeys(firstName);
     }
 
     private void inputLastName(String lastName) {
-        registerPage.waitUntilFindElement(registerPage.getLastNameInput());
         registerPage.getLastNameInput().clear();
         registerPage.getLastNameInput().sendKeys(lastName);
     }
 
     private void inputEmail(String email) {
-        registerPage.waitUntilFindElement(registerPage.getEmailInput());
         registerPage.getEmailInput().clear();
         registerPage.getEmailInput().sendKeys(email);
     }
 
     private void inputTelephone(String telephone) {
-        registerPage.waitUntilFindElement(registerPage.getTelephoneInput());
         registerPage.getTelephoneInput().clear();
         registerPage.getTelephoneInput().sendKeys(telephone);
     }
 
     private void inputPassword(String password) {
-        registerPage.waitUntilFindElement(registerPage.getPasswordInput());
         registerPage.getPasswordInput().clear();
         registerPage.getPasswordInput().sendKeys(password);
         registerPage.getPasswordConfirmInput().clear();
@@ -77,7 +72,6 @@ public class RegisterPageBL {
     }
 
     public void verifyUserRegistration() {
-        registerPage.waitUntilFindElement(successRegisterPage.getSuccessTitle());
         String expectedMessage = "Your Account Has Been Created!";
         Assert.assertEquals(successRegisterPage.getSuccessTitle().getText(), expectedMessage, "Incorrect page title");
     }
