+++ src/test/java/UserLoginTest.java
@@ -1,4 +1,3 @@
-import com.opencart.driver.DriverRepository;
 import com.opencart.navigation.Navigation;
 import com.opencart.steps.*;
 import org.testng.annotations.AfterMethod;
@@ -11,19 +10,8 @@ import static com.opencart.enums.URLs.BASE_URL;
 public class UserLoginTest extends BaseTest {
 
     @BeforeMethod
-    public void registerUserWithValidParameters() {
-        new Navigation().navigateToUrl(BASE_URL.getValue());
-        MainPageBL mainPageBL = new MainPageBL();
-        mainPageBL.getHeaderPageBL()
-                .clickOnMyAccountButton()
-                .clickOnRegisterButton()
-                .registerNewPerson()
-                .verifyUserRegistration();
-        mainPageBL.getHeaderPageBL()
-                .clickOnMyAccountButton()
-                .clickOnLogoutButton()
-                .clickContinue()
-                .verifyLogout();
+    public void registerUser() {
+        new MainPageBL().registerNewUser();
     }
 
     @Test
@@ -43,11 +31,7 @@ public class UserLoginTest extends BaseTest {
     }
 
     @AfterMethod
-    public void logout() {
-        new MainPageBL().getHeaderPageBL()
-                .clickOnMyAccountButton()
-                .clickOnLogoutButton()
-                .clickContinue()
-                .verifyLogout();
+    public void logoutOfUserProfile() {
+        new MainPageBL().logout();
     }
 }
