+++ src/test/java/UserRegisterTest.java
@@ -1,13 +1,7 @@
 import com.opencart.navigation.Navigation;
 import com.opencart.steps.MainPageBL;
-import com.opencart.steps.RegisterPageBL;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.Test;
-
-import java.io.FileNotFoundException;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.util.Properties;
 
 import static com.opencart.enums.URLs.BASE_URL;
 
@@ -25,10 +19,6 @@ public class UserRegisterTest extends BaseTest {
 
     @AfterMethod
     public void logout() {
-        new MainPageBL().getHeaderPageBL()
-                .clickOnMyAccountButton()
-                .clickOnLogoutButton()
-                .clickContinue()
-                .verifyLogout();
+        new MainPageBL().logout();
     }
 }
